Back/requirements.txt
# --- Web framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==11.0.3

# --- Base de datos ---
motor==3.6.0
beanie==1.26.0

# --- Seguridad ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
slowapi==0.1.9
pydantic[email]==2.9.0
pydantic-settings==2.5.0

# --- Solana ---
solana==0.34.0
solders==0.21.0

# --- IA / Gemini ---
google-generativeai==0.8.0

# --- Lógica difusa y bayesiana ---
scikit-fuzzy==0.4.2
pgmpy==0.1.26
numpy==1.26.4
pandas==2.2.3
scipy==1.13.1
scikit-learn==1.5.2
joblib==1.4.2

# --- Utils ---
python-multipart==0.0.12
aiofiles==24.1.0
httpx==0.27.0
pip install -r requirements.txt